todo-app/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.46
python-dotenv==0.19.0
gunicorn==20.1.0
